+++ compiler-tests/src/test/data/diagnostic/fir/dependencygraph/GraphsCannotDirectlyExtendOtherGraphs.kt
@@ -1,0 +1,14 @@
+// RENDER_DIAGNOSTICS_FULL_TEXT
+
+// MODULE: lib
+@DependencyGraph
+interface ParentGraph {
+  val value: Int
+
+  @Provides
+  fun provideInt(): Int = 3
+}
+
+// MODULE: main(lib)
+@DependencyGraph
+interface ChildGraph : <!DEPENDENCY_GRAPH_ERROR!>ParentGraph<!>